Refactor template result cells
