+++ intermine-ios/TemplateDetailTableViewController.swift
@@ -9,148 +9,6 @@
 import UIKit
 
 
-class LookupCell: UITableViewCell, UITextFieldDelegate {
-    
-    static let identifier = "LookupCell"
-    @IBOutlet weak var titleLabel: UILabel?
-    @IBOutlet weak var lookupTextField: UITextField?
-    var index: Int? = nil
-    
-    var query: TemplateQuery? {
-        didSet {
-            lookupTextField?.text = self.query?.getValue()
-        }
-    }
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        titleLabel?.text = "Lookup"
-        lookupTextField?.delegate = self
-        lookupTextField?.addTarget(self, action: #selector(textFieldEdited(_:)), for: UIControlEvents.editingChanged)
-    }
-    
-    func textFieldEdited(_ sender : UITextField) {
-        guard let index = self.index else {
-            return
-        }
-        var info: [String: Any] = [:]
-        if let value = sender.text {
-            info = ["value": value, "index": index]
-        } else {
-            info = ["value": "", "index": index]
-        }
-        NotificationCenter.default.post(name: NSNotification.Name(rawValue: Notifications.valueChanged), object: self, userInfo: info)
-    }
-}
-
-protocol OperationSelectionCellDelegate: class {
-    func operationSelectionCellDidTapSelectButton(cell: OperationSelectionCell)
-}
-
-class OperationSelectionCell: UITableViewCell, UITextFieldDelegate {
-    
-    // TODO: To refactor duplicated code into parent class
-    
-    static let identifier = "OperationSelectionCell"
-    @IBOutlet weak var operationLabel: UILabel?
-    @IBOutlet weak var selectOperationButton: UIButton?
-    @IBOutlet weak var valueTextField: UITextField?
-    
-    var index: Int? = nil
-    
-    weak var delegate: OperationSelectionCellDelegate?
-    
-    var query: TemplateQuery? {
-        didSet {
-            operationLabel?.text = self.query?.getOperation()
-            valueTextField?.text = self.query?.getValue()
-        }
-    }
-    
-    var operation: String? = "==" {
-        didSet {
-            operationLabel?.text = self.operation
-        }
-    }
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        selectOperationButton?.setTitle(String.localize("Templates.CTA.SelectOperation"), for: .normal)
-        selectOperationButton?.titleLabel?.lineBreakMode = NSLineBreakMode.byWordWrapping
-        selectOperationButton?.titleLabel?.numberOfLines = 2
-        selectOperationButton?.titleLabel?.textAlignment = NSTextAlignment.center
-        NotificationCenter.default.addObserver(self, selector: #selector(self.operationChanged(_:)), name: NSNotification.Name(rawValue: Notifications.operationChanged), object: nil)
-        valueTextField?.delegate = self
-        valueTextField?.addTarget(self, action: #selector(textFieldEdited(_:)), for: UIControlEvents.editingChanged)
-    }
-    
-    deinit {
-        NotificationCenter.default.removeObserver(self)
-    }
-    
-    func textFieldEdited(_ sender : UITextField) {
-        guard let index = self.index else {
-            return
-        }
-        var info: [String: Any] = [:]
-        if let value = sender.text {
-            info = ["value": value, "index": index]
-        } else {
-            info = ["value": "", "index": index]
-        }
-        NotificationCenter.default.post(name: NSNotification.Name(rawValue: Notifications.valueChanged), object: self, userInfo: info)
-    }
-    
-    @IBAction func operationSelectButtonTapped(_ sender: Any) {
-        self.delegate?.operationSelectionCellDidTapSelectButton(cell: self)
-    }
-    
-    // MARK: Notification when operation is changed
-    
-    func operationChanged(_ notification: NSNotification) {
-        if let op = notification.userInfo?["op"] as? String {
-            self.query?.changeOperation(operation: op)
-            operationLabel?.text = op
-        }
-    }
-}
-
-class DescriptionCell: UITableViewCell {
-    
-    static let identifier = "DescripitonCell"
-    @IBOutlet weak var descriptionLabel: UILabel?
-    
-    var info: String? = "" {
-        didSet {
-            descriptionLabel?.text = self.info
-        }
-    }
-    
-}
-
-protocol ActionCellDelegate: class {
-    func actionCellDidTapSearchButton(actionCell: ActionCell)
-}
-
-class ActionCell: UITableViewCell {
-    
-    static let identifier = "ActionCell"
-    @IBOutlet weak var searchButton: UIButton?
-    
-    weak var delegate: ActionCellDelegate?
-    
-    override func awakeFromNib() {
-        super.awakeFromNib()
-        searchButton?.setTitle(String.localize("Templates.CTA.Search"), for: .normal)
-    }
-    
-    @IBAction func searchButtonTapped(_ sender: Any) {
-        self.delegate?.actionCellDidTapSearchButton(actionCell: self)
-    }
-    
-}
-
-
 class TemplateDetailTableViewController: UITableViewController, OperationSelectionCellDelegate, OperationSelectViewControllerDelegate, ActionCellDelegate {
     
     private var sortedQueries: [TemplateQuery] = []
